// frontends/main/src/app/core/model/operator.ts
/**
 * Rodano API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type Operator = 'EQUALS' | 'NOT_EQUALS' | 'CONTAINS' | 'NOT_CONTAINS' | 'GREATER' | 'GREATER_EQUALS' | 'LOWER' | 'LOWER_EQUALS' | 'NULL' | 'NOT_NULL' | 'BLANK' | 'NOT_BLANK';

export const Operator = {
    EQUALS: 'EQUALS' as Operator,
    NOT_EQUALS: 'NOT_EQUALS' as Operator,
    CONTAINS: 'CONTAINS' as Operator,
    NOT_CONTAINS: 'NOT_CONTAINS' as Operator,
    GREATER: 'GREATER' as Operator,
    GREATER_EQUALS: 'GREATER_EQUALS' as Operator,
    LOWER: 'LOWER' as Operator,
    LOWER_EQUALS: 'LOWER_EQUALS' as Operator,
    NULL: 'NULL' as Operator,
    NOT_NULL: 'NOT_NULL' as Operator,
    BLANK: 'BLANK' as Operator,
    NOT_BLANK: 'NOT_BLANK' as Operator
};